Generate all matrices in double precision floating point, and only
convert to single precision when sending matrices and vectors to the
GPU. Decrease floating point error by doing the intermediate
calculations at high precision.

+++ camera.h
@@ -9,19 +9,19 @@
 struct camera {
 
 	// Tilt from vertical, in radians:
-	float tilt;
+	double tilt;
 
 	// Rotation along z axis, in radians:
-	float rotate;
+	double rotate;
 
 	// Distance from camera to origin (camera height):
-	float distance;
+	double distance;
 
 	// Horizontal view angle in radians:
-	float view_angle;
+	double view_angle;
 
 	// Viewport aspect ratio:
-	float aspect_ratio;
+	double aspect_ratio;
 
 	// Zoom level:
 	int zoom;
@@ -33,21 +33,21 @@ struct camera {
 
 	// Various view and projection matrices:
 	struct {
-		float tilt[16];
-		float rotate[16];
-		float translate[16];
-		float radius[16];
-		float proj[16];
-		float view[16];
+		double tilt[16];
+		double rotate[16];
+		double translate[16];
+		double radius[16];
+		double proj[16];
+		double view[16];
 	} matrix;
 
 	// Inverse matrices:
 	struct {
-		float tilt[16];
-		float rotate[16];
-		float translate[16];
-		float radius[16];
-		float view[16];
+		double tilt[16];
+		double rotate[16];
+		double translate[16];
+		double radius[16];
+		double view[16];
 	} invert;
 
 	// Matrix update flags:
@@ -61,8 +61,8 @@ extern const struct camera *camera_get (void);
 extern void camera_updated_reset (void);
 extern void camera_unproject (union vec *, union vec *, const unsigned int x, const unsigned int y, const struct viewport *vp);
 extern void camera_set_aspect_ratio (const struct viewport *vp);
-extern void camera_set_view_angle (const float radians);
-extern void camera_set_rotate (const float radians);
-extern void camera_set_tilt (const float radians);
-extern void camera_set_distance (const float distance);
+extern void camera_set_view_angle (const double radians);
+extern void camera_set_rotate (const double radians);
+extern void camera_set_tilt (const double radians);
+extern void camera_set_distance (const double distance);
 extern bool camera_init (const struct viewport *vp);
